service增加StockService,用于书店添加新书 完成service层剩余测试类 cuj 18.1.12

+++ src/main/java/cuj/settlementsystem/service/BookStoreService.java
@@ -6,8 +6,12 @@ import cuj.settlementsystem.domain.Book;
  * Created by cujamin on 2018/1/11.
  */
 public interface BookStoreService {
+
     void showStockMap();
+
     void showBookInfo(String bookName);
+
     double checkPriceByBookName(String bookName);
-    Book buyBookByName(String bookName , float price);
+
+    Book buyBookByName(String bookName , double price);
 }
